上传了前端小程序

+++ src/main/java/work/likecu/share/model/UserMessage.java
@@ -16,7 +16,7 @@ public class UserMessage {
     private String userNickname;
 
     @Column(name = "`user_gender`")
-    private Byte userGender;
+    private Integer userGender;
 
     @Column(name = "`user_token`")
     private String userToken;
